Renaming file and adding to do items
