+++ src/apps/Art/types/genesisBlocks.ts
