+++ src/apps/Art/types/index.ts
@@ -1,6 +1,6 @@
 import {Artwork, ArtworkAttributes, Artworks} from 'apps/Art/types/artworks';
-import {ArtworkIdPayload, GenesisBlock, GenesisBlockPayload, UnsignedGenesisBlock} from 'apps/Art/types/blocks';
 import {ArtElectronStore} from 'apps/Art/types/electronStore';
+import {ArtworkIdPayload, GenesisBlock, GenesisBlockPayload, UnsignedGenesisBlock} from 'apps/Art/types/genesisBlocks';
 import {Manager, Page} from 'apps/Art/types/manager';
 import {QueuedBlock, QueuedBlockPayload} from 'apps/Art/types/queuedBlocks';
